+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.17] - 2026-04-30
+
+### Docs
+- Update README.md
+- Update TODO.md
+
+### Other
+- Update deta/cli.py
+- Update deta/formatter/graph.py
+- Update deta/monitor/prober.py
+- Update deta/web/app.py
+
 ## [0.2.16] - 2026-04-30
 
 ### Docs